Improve Codes By Fixing Narrowing Conversion

+++ Examples/StandAloneSensors.hpp
@@ -37,7 +37,7 @@ namespace airlib
                 output_stream << output.angular_velocity.x() << "\t" << output.angular_velocity.y() << "\t" << output.angular_velocity.z() << "\t";
                 output_stream << output.linear_acceleration.x() << "\t" << output.linear_acceleration.y() << "\t" << output.linear_acceleration.z() << "\n";
 
-                std::this_thread::sleep_for(std::chrono::duration<double>(period - (Utils::getTimeSinceEpochSecs() - last)));
+                std::this_thread::sleep_for(std::chrono::duration<double>(static_cast<double>(period) - (Utils::getTimeSinceEpochSecs() - last)));
 
                 last = Utils::getTimeSinceEpochSecs();
                 environment.update();
@@ -67,7 +67,7 @@ namespace airlib
                 output_stream << Utils::getTimeSinceEpochSecs() << "\t";
                 output_stream << output.pressure << "\t" << output.altitude << std::endl;
 
-                std::this_thread::sleep_for(std::chrono::duration<double>(period - (Utils::getTimeSinceEpochSecs() - last)));
+                std::this_thread::sleep_for(std::chrono::duration<double>(static_cast<double>(period) - (Utils::getTimeSinceEpochSecs() - last)));
 
                 last = Utils::getTimeSinceEpochSecs();
                 environment.update();
@@ -99,7 +99,7 @@ namespace airlib
                     output_stream << Utils::getTimeSinceEpochSecs() << "\t";
                     output_stream << output.pressure << "\t" << output.altitude << "\t" << environment.getState().geo_point.altitude << std::endl;
 
-                    std::this_thread::sleep_for(std::chrono::duration<double>(period - (Utils::getTimeSinceEpochSecs() - last)));
+                    std::this_thread::sleep_for(std::chrono::duration<double>(static_cast<double>(period) - (Utils::getTimeSinceEpochSecs() - last)));
 
                     last = Utils::getTimeSinceEpochSecs();
                     environment.update();
@@ -147,7 +147,7 @@ namespace airlib
                     output_stream << output.magnetic_field_body.x() << "\t" << output.magnetic_field_body.y() << "\t" << output.magnetic_field_body.z();
                     output_stream << std::endl;
 
-                    std::this_thread::sleep_for(std::chrono::duration<double>(period - (Utils::getTimeSinceEpochSecs() - last)));
+                    std::this_thread::sleep_for(std::chrono::duration<double>(static_cast<double>(period) - (Utils::getTimeSinceEpochSecs() - last)));
                     last = Utils::getTimeSinceEpochSecs();
                     environment.update();
                     mag.update();
@@ -196,7 +196,7 @@ namespace airlib
                             output_stream << "\t" << kinematics.pose.orientation.w() << "\t" << kinematics.pose.orientation.x() << "\t" << kinematics.pose.orientation.y() << "\t" << kinematics.pose.orientation.z();
                             output_stream << std::endl;
 
-                            std::this_thread::sleep_for(std::chrono::duration<double>(period - (Utils::getTimeSinceEpochSecs() - last)));
+                            std::this_thread::sleep_for(std::chrono::duration<double>(static_cast<double>(period) - (Utils::getTimeSinceEpochSecs() - last)));
                             last = Utils::getTimeSinceEpochSecs();
                             environment.update();
                             mag.update();
